+++ kinto_signer/updater.py
@@ -1,7 +1,24 @@
-from kinto.core.utils import COMPARISON
-from kinto_signer.serializer import canonical_json
+from kinto.core.events import ACTIONS
 from kinto.core.storage import Filter
 from kinto.core.storage.exceptions import UnicityError, RecordNotFoundError
+from kinto.core.utils import COMPARISON, build_request
+from kinto_signer.serializer import canonical_json
+
+
+def notify_resource_event(request, request_options, matchdict,
+                          resource_name, parent_id, record, action, old=None):
+    """Private helper that triggers resource events when the updater modifies
+    the source and destination objects.
+    """
+    fakerequest = build_request(request, request_options)
+    fakerequest.matchdict = matchdict
+    fakerequest.bound_data = request.bound_data
+    fakerequest.current_resource_name = resource_name
+    fakerequest.notify_resource_event(parent_id=parent_id,
+                                      timestamp=record['last_modified'],
+                                      data=record,
+                                      action=action,
+                                      old=old)
 
 
 class LocalUpdater(object):
@@ -43,17 +60,18 @@ class LocalUpdater(object):
 
         # Define resource IDs.
 
-        self.destination_bucket_id = '/buckets/%s' % self.destination['bucket']
-        self.destination_collection_id = '/buckets/%s/collections/%s' % (
+        self.destination_bucket_uri = '/buckets/%s' % (
+            self.destination['bucket'])
+        self.destination_collection_uri = '/buckets/%s/collections/%s' % (
             self.destination['bucket'],
             self.destination['collection'])
 
-        self.source_bucket_id = '/buckets/%s' % self.source['bucket']
-        self.source_collection_id = '/buckets/%s/collections/%s' % (
+        self.source_bucket_uri = '/buckets/%s' % self.source['bucket']
+        self.source_collection_uri = '/buckets/%s/collections/%s' % (
             self.source['bucket'],
             self.source['collection'])
 
-    def sign_and_update_destination(self):
+    def sign_and_update_destination(self, request):
         """Sign the specified collection.
 
         0. Create the destination bucket / collection
@@ -64,34 +82,67 @@ class LocalUpdater(object):
            server
         5. Send the signature to the Authoritative server.
         """
-        self.create_destination()
+        before = len(request.get_resource_events())
+
+        self.create_destination(request)
         records, last_modified = self.get_source_records()
         serialized_records = canonical_json(records, last_modified)
         signature = self.signer.sign(serialized_records)
 
-        self.push_records_to_destination()
-        self.set_destination_signature(signature)
-        self.update_source_status("signed")
+        self.push_records_to_destination(request)
+        self.set_destination_signature(signature, request)
+        self.update_source_status("signed", request)
 
-    def _ensure_resource_exists(self, resource_type, parent_id, record_id):
+        # Re-trigger events from event listener \o/
+        for event in request.get_resource_events()[before:]:
+            request.registry.notify(event)
+
+    def _ensure_resource_exists(self, resource_type, parent_id,
+                                record_id, request):
         try:
-            self.storage.create(
+            created = self.storage.create(
                 collection_id=resource_type,
                 parent_id=parent_id,
                 record={'id': record_id})
         except UnicityError:
-            pass
+            created = None
+        return created
 
-    def create_destination(self):
+    def create_destination(self, request):
         # Create the destination bucket/collection if they don't already exist.
         bucket_name = self.destination['bucket']
         collection_name = self.destination['collection']
 
-        self._ensure_resource_exists('bucket', '', bucket_name)
-        self._ensure_resource_exists(
+        created = self._ensure_resource_exists('bucket', '',
+                                               bucket_name,
+                                               request)
+        if created:
+            notify_resource_event(request,
+                                  {'method': 'PUT',
+                                   'path': self.destination_bucket_uri},
+                                  matchdict={'id': self.destination['bucket']},
+                                  resource_name="bucket",
+                                  parent_id='',
+                                  record=created,
+                                  action=ACTIONS.CREATE)
+
+        created = self._ensure_resource_exists(
             'collection',
-            self.destination_bucket_id,
-            collection_name)
+            self.destination_bucket_uri,
+            collection_name,
+            request)
+        if created:
+            notify_resource_event(request,
+                                  {'method': 'PUT',
+                                   'path': self.destination_collection_uri},
+                                  matchdict={
+                                      'bucket_id': self.destination['bucket'],
+                                      'id': self.destination['collection']
+                                  },
+                                  resource_name="collection",
+                                  parent_id=self.destination_bucket_uri,
+                                  record=created,
+                                  action=ACTIONS.CREATE)
 
         # Set the permissions on the destination collection.
         # With the current implementation, the destination is not writable by
@@ -99,7 +150,7 @@ class LocalUpdater(object):
         # https://github.com/Kinto/kinto-signer/issues/55
         permissions = {'read': ("system.Everyone",)}
         self.permission.replace_object_permissions(
-            self.destination_collection_id, permissions)
+            self.destination_collection_uri, permissions)
 
     def get_source_records(self, last_modified=None, include_deleted=False):
         # If last_modified was specified, only retrieve items since then.
@@ -136,7 +187,7 @@ class LocalUpdater(object):
 
         return collection_timestamp, records_count
 
-    def push_records_to_destination(self):
+    def push_records_to_destination(self, request):
         last_modified, records_count = self.get_destination_last_modified()
         if records_count == 0:
             last_modified = None
@@ -146,10 +197,11 @@ class LocalUpdater(object):
 
         # Update the destination collection.
         for record in new_records:
-            if record.get('deleted', False):
+            deleted = record.get('deleted', False)
+            if deleted:
                 try:
-                    self.storage.delete(
-                        parent_id=self.destination_collection_id,
+                    pushed = self.storage.delete(
+                        parent_id=self.destination_collection_uri,
                         collection_id='record',
                         object_id=record['id'],
                         last_modified=record['last_modified']
@@ -157,15 +209,33 @@ class LocalUpdater(object):
                 except RecordNotFoundError:
                     # If the record doesn't exists in the destination
                     # we are good and can ignore it.
-                    pass
+                    continue
             else:
-                self.storage.update(
-                    parent_id=self.destination_collection_id,
+                pushed = self.storage.update(
+                    parent_id=self.destination_collection_uri,
                     collection_id='record',
                     object_id=record['id'],
                     record=record)
 
-    def set_destination_signature(self, signature):
+            matchdict = {
+                'bucket_id': self.destination['bucket'],
+                'collection_id': self.destination['collection'],
+                'id': record['id']
+            }
+            record_uri = ('/buckets/{bucket_id}'
+                          '/collections/{collection_id}'
+                          '/records/{id}'.format(**matchdict))
+            notify_resource_event(
+                request,
+                {'method': 'DELETE' if deleted else 'PUT',
+                 'path': record_uri},
+                matchdict=matchdict,
+                resource_name="record",
+                parent_id=self.destination_collection_uri,
+                record=pushed,
+                action=ACTIONS.DELETE if deleted else ACTIONS.UPDATE)
+
+    def set_destination_signature(self, signature, request):
         # Push the new signature to the destination collection.
         parent_id = '/buckets/%s' % self.destination['bucket']
         collection_id = 'collection'
@@ -176,16 +246,32 @@ class LocalUpdater(object):
             object_id=self.destination['collection'])
 
         # Update the collection_record
-        del collection_record['last_modified']
-        collection_record['signature'] = signature
+        new_collection = dict(**collection_record)
+        new_collection.pop('last_modified', None)
+        new_collection['signature'] = signature
 
-        self.storage.update(
+        updated = self.storage.update(
             parent_id=parent_id,
             collection_id=collection_id,
             object_id=self.destination['collection'],
-            record=collection_record)
+            record=new_collection)
 
-    def update_source_status(self, status):
+        matchdict = dict(bucket_id=self.destination['bucket'],
+                         id=self.destination['collection'])
+        notify_resource_event(
+            request,
+            {
+                'method': 'PUT',
+                'path': self.destination_collection_uri
+            },
+            matchdict=matchdict,
+            resource_name="collection",
+            parent_id=self.destination_bucket_uri,
+            record=updated,
+            action=ACTIONS.UPDATE,
+            old=collection_record)
+
+    def update_source_status(self, status, request):
         parent_id = '/buckets/%s' % self.source['bucket']
         collection_id = 'collection'
 
@@ -195,11 +281,27 @@ class LocalUpdater(object):
             object_id=self.source['collection'])
 
         # Update the collection_record
-        del collection_record['last_modified']
-        collection_record['status'] = status
+        new_collection = dict(**collection_record)
+        new_collection.pop('last_modified', None)
+        new_collection['status'] = status
 
-        self.storage.update(
+        updated = self.storage.update(
             parent_id=parent_id,
             collection_id=collection_id,
             object_id=self.source['collection'],
-            record=collection_record)
+            record=new_collection)
+
+        matchdict = dict(bucket_id=self.source['bucket'],
+                         id=self.source['collection'])
+        notify_resource_event(
+            request,
+            {
+                'method': 'PUT',
+                'path': self.source_collection_uri
+            },
+            matchdict=matchdict,
+            resource_name="collection",
+            parent_id=self.source_bucket_uri,
+            record=updated,
+            action=ACTIONS.UPDATE,
+            old=collection_record)